Move methods to time-utilities
